Update express samples
