+++ sdk/web-pubsub/web-pubsub/samples/v1/typescript/tsconfig.json
@@ -11,7 +11,5 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": [
-    "src/**.ts"
-  ]
+  "include": ["src/**.ts"]
 }
